remove old benchmarking scripts
These scripts were written for the publication of Schubert et al 2016,
were they were used to benchmark AdapterRemoval v2.1.3 against a
specific set of versions of other tools.

The scripts were included in the repo for the sake of reproducability
and for this reason they have not been updated since. There is little
reason to bundle the scripts with current versions of AdapterRemoval.