use isal for streaming output
This allows the dependency on libz to be dropped, but makes isa-l a
mandatory dependency. Since streaming output is only intended for
backwards compatibility in rare cases, the increased file size at
level 3 (+5-10%) is considered acceptable.

The --gzip-stream option was additionally removed in favor of simply
using --gzip-level to 3 or below to enable isa-l. This was motivated
by the observation that isa-l outperformed libdeflate at these levels.
The streaming implementation was addtionally made multi-threaded, using
an implementation inspired by isa-l's igzip.