pad sequence prior to SIMD alignment
This allows for the entire alignment to be carried out using the fasted
supported instruction set, instead of having to fall back to slower
methods.  A single shared buffer is used to contain the padded sequences,
which is reused between adapter pairs. Overall this reduces the runtime of
the alignment loops by as much as 50% and 25% when using AVX2 for PE and SE
alignments respectively.