partially revert 3c8c6410 and replace checks
Accept some pointless (PE)  alignments, but skip those that can easily
be determined to be pointless.

The check is also moved to the public functions, resulting in a small
increase in throughput due to not having to construct sequences for
skipped alignments.

`max_shift` is made unsigned to match the CLI